chore: 16378: Code review comments
