+++ cryptography/hedera-cryptography-altbn128/src/main/java/com/hedera/cryptography/altbn128/AltBn128Field.java
@@ -21,7 +21,7 @@ import com.hedera.cryptography.altbn128.facade.FieldFacade;
 import com.hedera.cryptography.pairings.api.Field;
 import com.hedera.cryptography.pairings.api.FieldElement;
 import com.hedera.cryptography.pairings.api.PairingFriendlyCurve;
-import com.hedera.cryptography.utils.BigIntegerUtils;
+import com.hedera.cryptography.utils.ByteArrayUtils;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.math.BigInteger;
 
@@ -75,7 +75,7 @@ public class AltBn128Field implements Field {
     @Override
     public FieldElement fromBigInteger(@NonNull final BigInteger bigInteger) {
         return new AltBn128FieldElement(
-                facade.fromBytes(BigIntegerUtils.toLittleEndianBytes(bigInteger, facade.size())), this);
+                facade.fromBytes(ByteArrayUtils.toLittleEndianBytes(bigInteger, facade.size())), this);
     }
 
     /**
